+++ indexer/src/file_upload.py
@@ -2,6 +2,7 @@ import os
 import shutil
 import logging
 import asyncio
+import tempfile
 from typing import List
 from fastapi import APIRouter, Form, UploadFile
 from fastapi.responses import JSONResponse
@@ -28,15 +29,6 @@ def cleanup_dir(path: str) -> None:
     os.makedirs(path, exist_ok=True)
 
 
-def save_files(path: str, safepath: str, files: List[UploadFile]) -> None:
-    cleanup_dir(path)
-    for file in files:
-        filepath = os.path.join(path, file.filename)
-        check_if_path_inside_allowed_path(safepath, filepath)
-        with open(filepath, "wb") as out_file:
-            out_file.write(file.file.read())
-
-
 def move_files(dest_dir: str, source_dir: str) -> None:
     cleanup_dir(dest_dir)
     for file in os.listdir(source_dir):
@@ -49,17 +41,26 @@ def move_files(dest_dir: str, source_dir: str) -> None:
 async def create_upload_files(
     directory: str, files: List[UploadFile], branch: str = Form()
 ):
+    """
+    A method to upload files in a certain directory
+    Args:
+        directory: Repository name
+        files: File list
+        branch: Branch name
+
+    Returns:
+        Upload status
+    """
     if directory not in indexes:
         return JSONResponse(f"{directory} not found!", status_code=404)
 
     reindex_dir = indexes.get(directory)
     path = os.path.join(settings.files_dir, directory, branch)
-    temp_path = os.path.join(settings.temp_dir, directory, branch)
     async with lock:
         try:
-            check_if_path_inside_allowed_path(settings.temp_dir, temp_path)
-            save_files(temp_path, settings.temp_dir, files)
-            move_files(path, temp_path)
+            with tempfile.TemporaryDirectory() as temp_path:
+                check_if_path_inside_allowed_path(settings.temp_dir, temp_path)
+                move_files(path, temp_path)
             logging.info(f"Uploaded {len(files)} files")
         except Exception as e:
             logging.exception(e)
@@ -69,5 +70,6 @@ async def create_upload_files(
             return JSONResponse("File uploaded, reindexing is done!")
         except Exception as e:
             return JSONResponse(
-                f"File uploaded, but error occurred during re-indexing: {e}", status_code=500
+                f"File uploaded, but error occurred during re-indexing: {e}",
+                status_code=500,
             )
